renames general view to general terminal equipment view
