+++ src/groups/RouteNetworkDiagram/GeneralTerminalEquipmentView/GeneralTerminalEquipmentView.scss
