+++ src/groups/RouteNetworkDiagram/GeneralTerminalEquipmentView/GeneralTerminalEquipmentViewGql.ts
