+++ src/groups/RouteNetworkDiagram/GeneralTerminalEquipmentView/index.ts
@@ -1,0 +1,4 @@
+import GeneralTerminalEquipmentView from "./GeneralTerminalEquipmentView";
+import "./GeneralTerminalEquipmentView.scss";
+
+export default GeneralTerminalEquipmentView;
