--- src/groups/RouteNetworkDiagram/GeneralView/index.ts
@@ -1,4 +1,0 @@
-import GeneralView from "./GeneralView";
-import "./GeneralView.scss";
-
-export default GeneralView;
